rename: 배포 관련 오류로 인한 폴더명 변경
