--- src/shared/index.ts
@@ -1,2 +1,0 @@
-export { GatheringCard } from '@/shared/ui/gatheringCard/GatheringCard';
-export { JobTag } from '@/shared/ui/jobTag/JobTag';
