+++ src/shared/ui/gatheringCard/JobTag.module.scss
