+++ src/shared/ui/gatheringCard/JobTag.tsx
