+++ src/shared/ui/gatheringCard/index.ts
@@ -1,0 +1,2 @@
+export { GatheringCard } from './GatheringCard';
+export { JobTag } from './JobTag';
