+++ src/shared/ui/index.ts
@@ -1,1 +1,2 @@
 export { Button } from './Button/Button';
+export * from './GatheringCard/index';
